function selector fix
