+++ constants/addresses.json
@@ -49,11 +49,11 @@
                         "0x175accdc",
                         "0x0f449d71",
                         "0x493189f0",
-                        "0x4210314d",
-                        "0x91ac6ca8",
-                        "0x0adedc02",
-                        "0x85f2aeda",
-                        "0x6b1ef56f",
+                        "0xcc713a04",
+                        "0x56a75868",
+                        "0x9fda64bd",
+                        "0xf497df75",
+                        "0x07ed2379",
                         "0x83800a8e",
                         "0x8770ba91",
                         "0x19367472",
@@ -72,6 +72,23 @@
                 {
                     "router": "0x6fd4383cB451173D5f9304F041C7BCBf27d561fF",
                     "selectors" : [
+                        "0xa76dfc3b",
+                        "0x89af926a",
+                        "0x188ac35d",
+                        "0x175accdc",
+                        "0x0f449d71",
+                        "0x493189f0",
+                        "0xcc713a04",
+                        "0x56a75868",
+                        "0x9fda64bd",
+                        "0xf497df75",
+                        "0x07ed2379",
+                        "0x83800a8e",
+                        "0x8770ba91",
+                        "0x19367472",
+                        "0xe2c95c82",
+                        "0xea76dddf",
+                        "0xf7a70056"
                     ]
                 }
             ]
@@ -84,9 +101,9 @@
                 {
                     "router": "0x6131B5fae19EA4f9D964eAc0408E4408b66337b5",
                     "selectors" : [
-                        "0x93eca257",
-                        "0x546da4a1",
-                        "0xc7306aaa"
+                        "0xe21fd0e9",
+                        "0x59e50fed",
+                        "0x8af033fb"
                     ]
                 }
             ]
